+++ src/app.module.ts
@@ -16,11 +16,13 @@ import {Transport} from '@nestjs/microservices';
 import {GameModule} from './game/game.module';
 import {MemberModule} from './member/member.module';
 import {EmpireModule} from './empire/empire.module';
-import { PresetsModule } from './presets/presets.module';
+import {PresetsModule} from './presets/presets.module';
 import {SystemModule} from "./system/system.module";
-import { GameLogicModule } from './game-logic/game-logic.module';
+import {GameLogicModule} from './game-logic/game-logic.module';
 import {IncomingMessage} from 'http';
 import {AuthService} from './auth/auth.service';
+import {FriendModule} from "./friend/friend.module";
+import {JobModule} from "./job/job.module";
 
 @Module({
   imports: [
@@ -52,10 +54,12 @@ import {AuthService} from './auth/auth.service';
     }),
     AuthModule,
     UserModule,
+    FriendModule,
     GameModule,
     MemberModule,
     SystemModule,
     EmpireModule,
+    JobModule,
     AchievementSummaryModule,
     AchievementModule,
     PresetsModule,