+++ src/game-logic/game-logic.module.ts
@@ -1,22 +1,19 @@
-import {Module} from '@nestjs/common';
+import {forwardRef, Module} from '@nestjs/common';
 import {GameLogicService} from './game-logic.service';
-import {environment} from '../environment';
-import {GameLogicScheduler} from './game-logic.scheduler';
-import {GameModule} from '../game/game.module';
 import {SystemModule} from '../system/system.module';
 import {EmpireModule} from '../empire/empire.module';
 import {GameLogicController} from './game-logic.controller';
-import {GameLogicHandler} from './game-logic.handler';
 import {MemberModule} from '../member/member.module';
 
 @Module({
   imports: [
-    GameModule,
+    forwardRef(() => require('../game/game.module').GameModule),
     MemberModule,
     SystemModule,
     EmpireModule,
   ],
-  providers: environment.passive ? [GameLogicService, GameLogicHandler] : [GameLogicService, GameLogicHandler, GameLogicScheduler],
+  providers: [GameLogicService],
+  exports: [GameLogicService],
   controllers: [GameLogicController],
 })
 export class GameLogicModule {