+++ src/user/user.module.ts
@@ -1,6 +1,5 @@
 import {forwardRef, Module} from '@nestjs/common';
 import {MongooseModule} from '@nestjs/mongoose';
-
 import {AuthModule} from '../auth/auth.module';
 import {environment} from '../environment';
 import {UserController} from './user.controller';